+++ pages/dev/src/payload.config.ts
@@ -1,4 +1,4 @@
-import { getPageUrl, payloadPagesPlugin } from '@jhb.software/payload-pages-plugin'
+import { payloadPagesPlugin } from '@jhb.software/payload-pages-plugin'
 import { mongooseAdapter } from '@payloadcms/db-mongodb'
 import path from 'path'
 import { buildConfig } from 'payload'
@@ -23,11 +23,6 @@ export default buildConfig({
       password: 'test',
     },
     user: 'users',
-    livePreview: {
-      // For testing purposes, we only want to live preview the pages collection
-      collections: ['pages'],
-      url: ({ data }) => getPageUrl({ path: data.path, preview: true })!,
-    },
   },
   collections: [
     Pages,
@@ -62,7 +57,14 @@ export default buildConfig({
   i18n: {
     supportedLanguages: { en, de },
   },
-  plugins: [payloadPagesPlugin({})],
+  plugins: [
+    payloadPagesPlugin({
+      generatePageURL: ({ path, preview }) =>
+        path && process.env.NEXT_PUBLIC_FRONTEND_URL
+          ? `${process.env.NEXT_PUBLIC_FRONTEND_URL}${preview ? '/preview' : ''}${path}`
+          : null,
+    }),
+  ],
   async onInit(payload) {
     const existingUsers = await payload.find({
       collection: 'users',