+++ pages/src/components/client/hooks/useCollectionConfig.ts
@@ -1,3 +1,5 @@
+'use client'
+
 import { useConfig, useDocumentInfo } from '@payloadcms/ui'
 import { ClientCollectionConfig } from 'payload'