+++ pages/src/components/client/hooks/usePageCollectionConfigAtrributes.ts
@@ -1,6 +1,8 @@
-import { asPageCollectionConfigOrThrow } from '../../../collections/PageCollectionConfig.js'
+'use client'
+
 import { PageCollectionConfigAttributes } from '../../../types/PageCollectionConfigAttributes.js'
 import { useCollectionConfig } from './useCollectionConfig.js'
+import { asPageCollectionConfigOrThrow } from '../../../utils/pageCollectionConfigHelpers.js'
 
 /**
  * Returns the PageCollectionConfigAttributes for the collection of the document.