+++ pages/src/hooks/selectDependentFieldsBeforeOperation.ts
@@ -1,7 +1,7 @@
 import { CollectionBeforeOperationHook } from 'payload'
-import { asPageCollectionConfigOrThrow } from '../collections/PageCollectionConfig.js'
+import { asPageCollectionConfigOrThrow } from '../utils/pageCollectionConfigHelpers.js'
 import { dependentFields } from './setVirtualFields.js'
-import { getSelectType } from '../utils/getSelectType.js'
+import { getSelectMode } from 'payload/shared'
 
 /**
  * A CollectionBeforeOperationHook that alters the select in case a virtual field is selected
@@ -13,16 +13,20 @@ export const selectDependentFieldsBeforeOperation: CollectionBeforeOperationHook
   operation,
   context,
 }) => {
+  // // TODO: args.id will be undefined for findMany operations but data.id will be defined in the afterOperation hook
+  // // this means for findMany operations, removing the originally unselected fields will not work
+  context.rootOperation = `${args.id}-${args.collection.config.slug}`
+
   // Make the select object available to the setVirtualFields hook by adding it to the context
   context.select = args.select
 
   if (operation == 'read' && args.select) {
     const originalSelect = args.select
     const pageConfig = asPageCollectionConfigOrThrow(args.collection.config)
-    const selectType = getSelectType(args.select)
+    const selectMode = getSelectMode(args.select)
     const dependendSelectedFields = dependentFields(pageConfig)
 
-    if (selectType === 'include') {
+    if (selectMode === 'include') {
       const selectedFields = Object.keys(args.select)
 
       const virtualFields = ['path', 'breadcrumbs', 'meta']
@@ -35,6 +39,22 @@ export const selectDependentFieldsBeforeOperation: CollectionBeforeOperationHook
           ...dependendSelectedFields.reduce((acc, field) => ({ ...acc, [field]: true }), {}),
         }
 
+        const parentFieldName = pageConfig.page.parent.name
+        const parentCollectionSlug = pageConfig.page.parent.collection
+
+        // when the user has not selected the parent field, but the parent field is added above, only populate the breadcrumbs field of it.
+        if (!originalSelect[parentFieldName]) {
+          args.populate = {
+            ...args.populate,
+
+            // only populate the breadcrumbs field of the parent field, otherwise, for every read operation, all fields of the parent would be unecessaryly returned
+            [parentCollectionSlug]: {
+              breadcrumbs: true,
+              ...args.populate?.[parentCollectionSlug],
+            },
+          }
+        }
+
         // Store the original select so that deleteUnselectedFieldsAfterRead can properly handle field exclusion
         context.originalSelect = originalSelect
 
@@ -44,7 +64,7 @@ export const selectDependentFieldsBeforeOperation: CollectionBeforeOperationHook
         // Indicate that the virtual fields should be generated in the setVirtualFields hook
         context.generateVirtualFields = true
       }
-    } else if (selectType === 'exclude') {
+    } else if (selectMode === 'exclude') {
       const selectedFields = Object.keys(args.select)
       const virtualFields = ['path', 'breadcrumbs', 'meta']
       const allVirtualFieldsDeselected = virtualFields.every((field) =>