+++ pages/src/translations/en.ts
@@ -13,8 +13,6 @@ export const en: GenericTranslationsObject = {
     alternatePath: 'Alternate Path',
     breadcrumbs: 'Breadcrumbs',
     breadcrumb: 'Breadcrumb',
-    saveDocumentToPreview: 'Save the document to preview the changes',
-    openWebsitePageInPreviewMode: 'Open Website Page in preview mode',
     syncSlugWithX: 'Sync with {X}',
     slugWasChangedFromXToY:
       'The slug was changed from <code>{X}</code> to <code>{Y}</code>. This requires a redirection from the old to the new page path to be manually created.',